fixed really fatal bug
there was no return statement in getSecretNum(). this lead to the bug "TypeError: 'NoneType' object is not subscriptable". basically the value of secretNum was None (because the method in which it sets a value did not return any value). I was able to verify the value of secretNum by printing it and commenting out the code that made the program crash.

Learning new things everyday, hooray!